Migration of resource server
* removed resource server code
* removed marketplace api
* removed marketplace api test
* made tests agnostic of realm
* added missing http headers
* fix on db schemes ( added default values on some tables).

+++ app/tests/OpenStackIDBaseTest.php
@@ -19,26 +19,40 @@ abstract class OpenStackIDBaseTest extends TestCase {
 
     protected function prepareForTests()
     {
-        DB::table('banned_ips')->delete();
-        DB::table('user_exceptions_trail')->delete();
-        DB::table('server_configuration')->delete();
-        DB::table('server_extensions')->delete();
-
-        DB::table('oauth2_client_api_scope')->delete();
-        DB::table('oauth2_client_authorized_uri')->delete();
-        DB::table('oauth2_access_token')->delete();
-        DB::table('oauth2_refresh_token')->delete();
-        DB::table('oauth2_client')->delete();
-
-        DB::table('openid_trusted_sites')->delete();
-        DB::table('openid_associations')->delete();
-        DB::table('openid_users')->delete();
-
-        DB::table('oauth2_api_endpoint_api_scope')->delete();
-        DB::table('oauth2_api_endpoint')->delete();
-        DB::table('oauth2_api_scope')->delete();
-        DB::table('oauth2_api')->delete();
-        DB::table('oauth2_resource_server')->delete();
+        if (Schema::hasTable('banned_ips'))
+            DB::table('banned_ips')->delete();
+        if (Schema::hasTable('user_exceptions_trail'))
+            DB::table('user_exceptions_trail')->delete();
+        if (Schema::hasTable('server_configuration'))
+            DB::table('server_configuration')->delete();
+        if (Schema::hasTable('server_extensions'))
+            DB::table('server_extensions')->delete();
+        if (Schema::hasTable('oauth2_client_api_scope'))
+            DB::table('oauth2_client_api_scope')->delete();
+        if (Schema::hasTable('oauth2_client_authorized_uri'))
+            DB::table('oauth2_client_authorized_uri')->delete();
+        if (Schema::hasTable('oauth2_access_token'))
+            DB::table('oauth2_access_token')->delete();
+        if (Schema::hasTable('oauth2_refresh_token'))
+            DB::table('oauth2_refresh_token')->delete();
+        if (Schema::hasTable('oauth2_client'))
+            DB::table('oauth2_client')->delete();
+        if (Schema::hasTable('openid_trusted_sites'))
+            DB::table('openid_trusted_sites')->delete();
+        if (Schema::hasTable('openid_associations'))
+            DB::table('openid_associations')->delete();
+        if (Schema::hasTable('openid_users'))
+            DB::table('openid_users')->delete();
+        if (Schema::hasTable('oauth2_api_endpoint_api_scope'))
+            DB::table('oauth2_api_endpoint_api_scope')->delete();
+        if (Schema::hasTable('oauth2_api_endpoint'))
+            DB::table('oauth2_api_endpoint')->delete();
+        if (Schema::hasTable('oauth2_api_scope'))
+            DB::table('oauth2_api_scope')->delete();
+        if (Schema::hasTable('oauth2_api'))
+            DB::table('oauth2_api')->delete();
+        if (Schema::hasTable('oauth2_resource_server'))
+            DB::table('oauth2_resource_server')->delete();
 
         parent::prepareForTests();
     }
